+++ NEWS.md
@@ -12,6 +12,7 @@
 * Defold plugin: Allow overriding z value also when exporting to .collection (#3214)
 * Qt 6: Fixed invisible tileset tabs when only a single tileset is open
 * Fixed positioning of point object name labels (by Logan Higinbotham, #3400)
+* Fixed slight drift when zooming the map view in/out
 * Fixed compile against Qt 6.4
 * snap: Added Wayland platform plugin and additional image format plugins
 * AppImage: Updated to Sentry 0.5.2