+++ NEWS.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Restored Tiled 1.8 file format compatibility by default (#3560)
 * Added action search popup on Ctrl+Shift+P (with dogboydog, #3449)
 * Added file system actions also for tileset image based tilesets (#3448)
 * Fixed new layer names to be always unique (by Logan Higinbotham, #3452)